Add Orbit Bridge farms to update-monitor.
